fixed an error in unlinking posts' images if they didn't exist

+++ routes/posts.js
@@ -41,9 +41,11 @@ router.delete("/:id", verifyToken, async (req, res) => {
     req.body.userId = req.user.id;
     const post = await Post.findById(req.params.id);
     if (post.userId === req.body.userId) {
-      await fs.unlink(`./public/images/${post.img}`, (err) => {
-        if (err) console.log(err);
-      });
+      if (post.img) {
+        await fs.unlink(`./public/images/${post.img}`, (err) => {
+          if (err) console.log(err);
+        });
+      }
       await post.deleteOne();
       res.status(200).json("the post has been deleted");
     } else {
